Change return type of rumqttc event loop (#136)
Previously, the event loop yielded items of type
Result<(Option<Incoming>, Option<Outgoing>), Err>.
However, it is never the case that an Incoming and an Outgoing
event are yielded at the same time, which makes handling
this result unnecessarily tricky. Therefore, the return type is
changed to Result<Event, Err> where Event is an enum containing
Incoming and Outgoing as variants.